add: asNode updater
