+++ packages/dataparcels/.size-limit.json
@@ -1,6 +1,6 @@
 [
     {
-        "limit": "10.2 KB",
+        "limit": "10.3 KB",
         "path": "lib/index.js"
     },
     {
@@ -20,12 +20,16 @@
         "path": "deleted.js"
     },
     {
-        "limit": "7.6 KB",
+        "limit": "7.7 KB",
         "path": "ParcelShape.js"
     },
     {
         "limit": "5.9 KB",
         "path": "ParcelNode.js"
+    },
+    {
+        "limit": "6 KB",
+        "path": "asNode.js"
     },
     {
         "limit": "6 KB",
