+++ packages/dataparcels/asNode.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('./lib/parcelNode/asNode.js').default;
