+++ packages/dataparcels/src/errors/Errors.js
@@ -9,3 +9,4 @@ export const ReducerInvalidActionError = (actionType: string) => new Error(`"${a
 export const ReducerInvalidStepError = (stepType: string) => new Error(`"${stepType}" is not a valid action step type`);
 export const ChangeRequestNoPrevDataError = () =>  new Error(`ChangeRequest data cannot be accessed before setting changeRequest.prevData`);
 export const ShapeUpdaterNonShapeChildError = () =>  new Error(`Every child value on a collection returned from a shape updater must be a ParcelShape`);
+export const AsNodeReturnNonParcelNodeError = () =>  new Error(`The return value of an asNode() updater must be a ParcelNode`);
