+++ packages/react-dataparcels/asNode.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('dataparcels/asNode.js');
